Implement banding reduction
This commit introduces a banding reduction algorithm. Basically, we
compute the average value of each line, then, for a particular line,
compare it to the average of the averages of nearbye lines. We
compute a correction factor. If it's positive, then it means the line
is darker than its neighbors, and we correct it. We perform several
passes.

This algorithm makes use of the incubating vector API.

Fixes #5