Add coronagraph image
This commit implements a new algorithm which is responsible for finding
the borders of the sun, then generate samples which are used to perform
ellipse fitting. We use ellipse fitting because it is very likely that
the input video will not generate a perfectly round sun. In fact, in
most cases it will be an ellipse, so this algorithm performs edge
detection, then computes the parameters of the ellipse.

For now, this ellipse is _only_ used to generate the virtual coronagraph
image, but it should be possible to use the same information to
fix the image geometry.

Closes #3