Refactor how images are emitted
Instead of relying on the `emitImage` method which mixes processing
with image writing, we now have a framework which allows performing
computations in parallel of writing images, while also allowing to
reduce the results from computations of previous steps. For example,
the ellipse fitting results will be usable for distortion correction.