Move all computations to 16-bit
This commit is a major refactoring of how image math is performed,
in order to compute everything with a minimum of 16-bit precision
(we use short for data, 32-bit float for math, 64-bit double for
when precision loss is annoying).

This is done because some input files will use 16-bit data, so if
we perform computations with only 8 bit, we are effectively reducing
resolution.

Note that this significantly changes the default coefficients, and
that they need to be recalibrated for different kind of videos.
This is going to be done in a subsequent commit.

As part of this work, a new debug window has been added, which
allows "seeing" the detected spectral line given some parameters.