Implement geometry correction
This commit introduces image geometry correction. Based on the
discovered ellipse parameters which describe the shape of the
solar disk, we perform a rotation of the disk and adjust the
image width and height so that the sun is round.

This currently breaks the banding correction algorithm since
it works after the transformation, and that it assumes vertical
artifacts.

Fixes #2
Fixes #4