Rotate image before processing
This commit performs a left rotation before starting the workflow.
The reason is that we reconstruct the image top-down, but the sun
is scanned either from east to west or from west to east. We can't
know which orientation it is, so we arbitrarily rotate to the left,
assuming that east will be left.

In addition, the ellipse fitting algorithm was slightly improved
to remove outliers, especially because of protuberances.