Refactor processing
In order to prepare for doppler image creation, this commit refactors
how processing is executed, so that we can concurrently build multiple
images at different pixel shifts. For now, it only uses one value, but
we will be able to compute multiple at once. A change in how we perform
image reconstruction was done: instead of computing a polynomial for
each frame, we now assume that the distortion is the same in all
frames, and therefore start with computing an average frame and use it
to determine the distortion polynomial.