Add doppler image generation
This commit introduces doppler image generation. To do this, we
reworked how the workflow is done, so that a workflow can access
the previous steps, which is necessary to generate a doppler
image: we need to generate reconstructed images with a negative
and positive shift, then recombine both into a color image.

Fixes #9