Add a "mixed" image
This commit introduce a new "mixed" image which combines the coronagraph
with the disk view in order to make the protuberances more visible. The
rotation algorithm has been changed to fill the missing pixels with 0
instead of the black point as it gives better results from a set of test
images.