Crop geometry corrected image
This commit introduces cropping of the geometry corrected image.
Because the image can be rotated, the dimensions of the geometry
corrected image can become significantly larger than the initial
one. Similarly, the center may be off, which can further increase
the target image size.

This commit performs a new fitting after correcting the image to
determine the disk center, then crops the image around the disk,
using a 20% margin.

This also improves the overall subsequent steps quality as the
computations are done on an image which is closer to the actual
disk.