Rework tilt detection
This commit completely reworks how tilt detection works, in order to be
closer to what INTI does. Instead of relying on the rotation angle found
by ellipse regression, we know use the ellipse equation to find 2 points
which should be aligned and compute their alignment angle.

Once this angle is computed, we now produce an _unrotated_ image, but
simply a shear transform which shifts pixels according to the tilt angle.

Fixes #14