Add custom file naming strategies
This commit adds the possibility to change how output files are written.
It is now possible to select different naming patterns, or create a
custom one given a number of predefined tokens.

This is a pre-requisite for a batch mode, where it would be more
convenient to write all files in a single directory.