Add a background removal function
This function can be called with `remove_bg(img;tolerance)`. It will use
the ellipse mask to determine pixels which are outside of the sun disk,
then compute the average of the image outside of the disk. It will then
attempt to remove the background while preserving areas which are close
to the limb.