Make it possible to re-execute scripts when image shifts are missing
After processing an image, we enable the image math editor on the right.
This editor allows working with image shifts which are already computed.
However, sometimes you want to try with other parameters, and it would
be useful to be able to recompute with new image shifts. This is what
this commit does, by automatically reprocessing the video if image
shifts are missing.