Fix io executor not blocking
There was a bug in the parallel executor, which made it that the blocking
call wasn't using the semaphore, resulting in too many tasks being executed
concurrently. In practice, it meant we could, in batch mode, read several
SER files in parallel when in practice we wanted to restrict this to a
single file.