Add batch mode processing scripts
This commit introduces a new section in ImageMath scripts,
the `[[batch]]` section, which works similarly to the main section,
but instead of working with images from a single SER file, uses
the outputs of the main section.

In other words, this section is executed at the end of a batch, and
allows working on outputs of individual SER files of the batch.
For example, if the main section declares a single output like this:

```
[outputs]
my_image=....
```

then in batch mode, it is possible to use all `my_image` and combine
them:

```
[[batch]]
[outputs]
stacked=median(my_image)
```

Therefore, while in the main section `my_image` represents a single
image, in the `[[batch]]` section it represents a list of images.