Add a spectral line profile chart
This commit introduces a new graph which shows the profile of the
spectral line being observed. The profile corresponds to the mean
intensity of the pixels in the average image, and is expressed
either in pixel shifts, or, if the binning and pixel size of the
camera have been provided, in wavelength in angstrom.