Add Richardson-Lucy deconvolution
This commit adds the ability to perform a Richardson-Lucy
deconvolution of images. The algorithm uses a synthetic PSF
which can be configured in terms of radius and sigma. By
default, deconvolution is disabled.

In order to make things clearer, image enhancement parameters
have been grouped into their own panel in the process
parameters window.