Introduce stacking and mosaic stitching
This commit introduce a couple of new features: image stacking
and mosaic composition. Both are exposed as functions in image
math scripts, respectively using the `stack` and `mosaic`
functions.

Stacking is based on an in-house algorithm which requires
the images to be roughly aligned and centered. It will determine,
using tiles, how each tile should move in order to match the
image with the best sharpness. A tile overlap is used in order
to create a sliding window which removes tiling artifacts. In
addition, if the overlap is large enough, the composition is
done using circles instead of squares, which further removes
visual artifacts.

The mosaicing algorithm uses a different technique, which
consists of warping images together to create a reference
image, which is then passed to the stacking algorithm.