Add new buttons for fitting images into window
The first button allows resetting the zoom level to the initial zoom,
which is made to fit the whole image into the window.

The 2d button will instead make sure that the solar disk fits into
the visible window, independently of how large is the image. This
button can be useful when images have large margins around the sun.