Fix polynomial detection
This commit fixes the polynomial detection for some edge cases. It appears
that in some videos, the first and last lines are too dark and can be
confused with the actual line that we want to detect. To avoid this, we're
filtering the lines so that they are not too close to the edges.

The average image computation is also modified in order to keep only the
brightest frames. This is done by sampling frames in the video file in
the beginning, and this is more accurate than the initial detection
algorithm.

In addition this fixes a bug where the polynomial regression algorithm
could return an empty array, in which case it means that detection failed.

Fixes #263