Fix race condition when reconstructing images
This commit fixes a race condition which could lead to incorrectly
reconstructed images. A typical symptom was that from one run to
the other the image was slightly different. In some rare cases
where the image was having a large tilt, this could even lead
to completely broken geometry correction.