Add automatic helium line processing
This commit introduces automatic helium line processing: if the Sodium D2
or Iron (Fe 1) line is detected, then we will automatically discover the
Helium line and produce an image with default settings. If the user isn't
heppy with the result, there's till the option to generate images using
ImageMath instead.