Improve speed detection
This commit reworks speed detection to be more accurate. It should result in less
false positives, while more accurately representing the areas where speeds are
detected.

The speed images are assigned a unique id, so that it is easier to locate them
and match images together.

In addition, the spectrum debugger will now show detected shifts, and the generated
debug images will tell where shifts are detected.