Add ability to perform manual line detection
This commit adds the ability to enter a custom polynomial. The
easiest way to get the polynomial coefficients is to use the
video analyzer, then press CTRL and click multiple times until
you have enough measurements. Once you click on the "compute"
button, the points will be fitted to a 3rd order polynomial
and reported automatically in the process parameters.

Fixes #348