Use 2-dimensional arrays for images
This commit changes the internal data structure used for images.
So far, it was using a 1d-array, which was optimal in terms of
memory consumption, but not very "user friendly" in the sense
that it required passing the width to many methods in order to
interpret the data.

With the idea of introducing Python scripting in the future,
there will be a need for lower-level data access, and it's
going to be easier for users to reason about with a 2d array.