Implement new stacking algorithm
This new algorithm was initially designed for the Sunscan, with the
help of Christian Buil. It has been integrated into JSol'Ex, making
it possible to use in Sunscan via the new scripting module, or it
may be implemented separately in Python, no decision has been made
yet.

However, the results are quite good with JSol'Ex, good enough to
replace the old stacking algorithm without noticeable quality
degradation.