Add sunspot detection
This commit introduces a new feature : detection of sunspots.
The detection is based on statistical analysis of spectrum
frames. Once detected, we can generate an overlay on different
images.

By default, the overlay will be generated on top of the
continuum image, as well as the "base" image. It is however
possible to generate more images via scripts.