Introduce image selector
This adds an interface to perform a review of the images generated
in batch mode. The UI allows the user to reject, approve or mark
as best an image. In case the batch processing is using scripts,
then the second part of the script, the [[batch]] section, will
only be applied on selected images.

Fixes #506