Make SER pixel depth recognition more robust
Before this change, we were looking for the maximum pixel depth only
around the middle of the SER file. While this worked for most cases,
it didn't when, for example, the SER file consisted of a long black
range of frames, then the sun only in the end (or the other way
around). Now sampling is regular around the whole SER file, we take
a look at 10% of the frames.