Improve script engine
This commit is a massive rewrite of the internal ImageMath script
engine. Before this change, the parser was hand written and hard
to maintain. We are now using a grammar written using CongoCC,
which generates an AST that we can use in the evaluation engine.

The new parser supports multiline strings, as well as multiline
expressions. It also adds support for including (statically)
other scripts, as well as declaring custom functions.

The grammar is ready for named arguments, but this isn't
implemented yet.