Rework notification mechanism
This commit reworks the progress events in order to be hierarchical,
instead of relying on a single notification event. Before, because
we were running tasks in parallel, a thread could hide the progress
bar because a task was completed, while other tasks were still
running. With the new mechanism, there's a best effort to restore
progress of other running tasks.

Fixes #519