Support for named parameters
This commit reworks the script engine in order to support named
arguments. It is now possible to use either a list of parameters,
like before, when invoking a function (e.g `img(0)`) or to invoke
the function using named arguments (e.g `img(ps: 0)`).

Named arguments are ordered, which makes it possible to map
the list of parameters to their named couterparts.