BASS2000 submission wizard
This commit introduces the possibility to submit images to the BASS2000
online databse directly in JSol'Ex. The submission process is only
available in single processing mode. This shouldn't be an issue since
it is only allowed to submit a single image per day, per spectral line.

The submission process goes through a wizard, which provides the following
features:

- BASS2000 introduction and guidelines
- orientation of image using GONG reference image
- tuning of the tilt using a blink image
- verification of submitted data using a validating form
- optional upload directly from the application

For now the wizard is disabled and must not be used unless approved by
the BASS2000 team.