Add automatic scripts
This commit adds the ability to automatically run a script for a particular
wavelength. The scripts are attached to a spectral line in the spectral line
editor. If the file being processed is using a spectral line which has automatic
scripts AND that the automatic scripts checkbox is checked, then the script
will automatically be executed.

Fixes #651