Add blend mode for BASS2000 image comparison
This commit adds a "blend" mode for easier BASS2000 image comparison. This mode
can be used alternatively with the "blink" mode, which gives the best of the 2
worlds.

In addition this commit introduces a sanity check for properties files, to make
sure no encoding errors are committed and fixes missing keys.