Add support for script parameters in UI
This commit introduces a new, optional section in scripts, which
is called `meta`. This section contains description about the
script, and more specifically, information about the parameters
that the script accepts. These parameters are different from
regular variables in the sense that they are exposed via the UI.

A script author can therefore write a "plugin" for JSol'Ex by
exposing a script with parameter definition, in which case the
users wouldn't have to understand where to modify the code: the
UI will display both parameters and tooltips.