Introduce stack_dedis function
This function lets the user stack images which are the result of
a DEDISTORT call. The distorsion map is used as an evaluation of
the quality of each image and instead of doing an average of all
images, we perform a weighted average.

The function supports providing a ratio of images to keep. In
addition, it is possible to use local distorsion, in which case
the algorithm will analyze the quality of each image around each
point. It means that in practice different regions of the disk
may use different images for stacking.