Add 3d profile and spectral cube
This commit introduces 2 new views on the profile tab. The first
one is a 3d view of the profile, which is based on the average
image. It can help diagnosing vignetting or slit defects.

The second one is a full spectral cube: it allows seeing the
intensity of particular wavelengths around the disk, as well
as seeing the variation of intensity of different wavelengths
at some point on the slit.