Add a 3d tomography rendering
This commit adds a new button in the profile tab, renamed "graphs" tab,
which allows generating a solar tomography. It works by generating 5
images at different pixel shifts, then projects them onto a sphere.

The renderers are responsible for blending the images together to
reveal the different heights in the chromosphere. There are 2 renderers,
depending on the machine capabilities (one using shaders, the other
using alpha blending). The shader one is more accurate and implements
a light propagation model.